requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
